modified datasource to be more cacha-friendly. you never know when you datasource will fail! better be prepared

+++ domain/src/commonMain/kotlin/com/github/lamba92/dragalialost/domain/entities/support/Skill.kt
@@ -3,7 +3,7 @@ package com.github.lamba92.dragalialost.domain.entities.support
 
 data class Skill(
     val name: String,
-    val skillPointCost: Int,
+    val skillPointCost: Long,
     val icon: String,
     val level1: SkillLevelData,
     val level2: SkillLevelData,
